+++ app/src/main/java/com/example/taskyapplication/auth/presentation/AuthViewModel.kt
@@ -1,7 +1,8 @@
-package com.example.taskyapplication.auth.domain
+package com.example.taskyapplication.auth.presentation
 
 import androidx.lifecycle.ViewModel
-import com.example.taskyapplication.TaskyAppPreferences
+import com.example.taskyapplication.auth.data.TaskyAppPreferences
+import com.example.taskyapplication.auth.domain.AuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
